+++ hfut_stu_lib/value.py
@@ -9,7 +9,8 @@ HF = 'HF'
 XC = 'XC'
 HOSTS = {HF: 'http://bkjw.hfut.edu.cn/', XC: 'http://222.195.8.201/'}
 
-TERM_PATTERN = re.compile(r'(\d{4})-\d{4}学年\s*第(一|二|二/三)学期(|/暑期)')
+# https://docs.python.org/3/library/re.html
+TERM_PATTERN = re.compile(r'(\d{4})(?:|学年)-\d{4}学年\s*第(一|二|二/三)学期(|/暑期)')
 ACCOUNT_PATTERN = re.compile(r'^\d{10}$')
 XC_PASSWORD_PATTERN = re.compile(r'^[\da-z]{6,12}$')
 HF_PASSWORD_PATTERN = re.compile(r'^\S{6,16}$')